+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Added Cineworld Dublin and its Irish film listings to the cinema planner.
+
 ## [1.6.9] - 2026-09-22
 
 ### Changed